requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
pydantic-settings
psycopg2-binary
passlib[bcrypt]
bcrypt
email-validator
python-jose[cryptography]
twilio
supabase
